% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-neighborhood-functions.R
\name{neighbor_mean}
\alias{neighbor_mean}
\title{for each cell, get the mean of x's values over its neighbors:}
\usage{
neighbor_mean(x, neighbors)
}
\arguments{
\item{x}{A numeric vector}

\item{neighbors}{A (probably sparse) adjacency matrix}
}
\description{
for each cell, get the mean of x's values over its neighbors:
}
