% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-neighborhood-functions.R
\name{neighbor_meantabulate}
\alias{neighbor_meantabulate}
\title{for each cell, tabulate the distinct values of x over its neighbors:}
\usage{
neighbor_meantabulate(x, neighbors)
}
\arguments{
\item{x}{A vector of categorical values}

\item{neighbors}{A (probably sparse) adjacency matrix}
}
\description{
for each cell, tabulate the distinct values of x over its neighbors:
}
