% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-neighborhood-functions.R
\name{subsampleNeighborsByRow}
\alias{subsampleNeighborsByRow}
\title{Subsample each row of a nearest neighbors matrix by 50\%,
assuming each row has exactly the same number of non-zero entries}
\usage{
subsampleNeighborsByRow(neighbors, p = 0.5)
}
\arguments{
\item{neighbors}{sparse matrix of neighbor relationships}

\item{p}{Subsampling rate, with 1 meaning all neighbors are kept.}
}
\value{
The same neighbors matrix, with p proportion of the entries in each row set to 0.
}
\description{
Subsample each row of a nearest neighbors matrix by 50\%,
assuming each row has exactly the same number of non-zero entries
}
